Relocating the accessTokenManager.
