+++ src/automowerPlatform.ts
@@ -6,7 +6,7 @@ import { SettingsEvent, StatusEvent } from './events';
 import { AutomowerAccessoryFactory, AutomowerAccessoryFactoryImpl } from './primitives/automowerAccessoryFactory';
 import { Localization, Y18nLocalization } from './primitives/localization';
 import { PlatformContainer, PlatformContainerImpl } from './primitives/platformContainer';
-import { AccessTokenManager, AccessTokenManagerImpl } from './services/automower/accessTokenManager';
+import { AccessTokenManager, AccessTokenManagerImpl } from './services/husqvarna/accessTokenManager';
 import { DiscoveryService, DiscoveryServiceImpl } from './services/automower/discoveryService';
 import { EventStreamService, EventStreamServiceImpl } from './services/automower/eventStreamService';
 import { PLATFORM_NAME, PLUGIN_ID } from './settings';
