+++ test/automowerPlatformSpy.ts
@@ -1,5 +1,5 @@
 import { AutomowerPlatform } from '../src/automowerPlatform';
-import { AccessTokenManager } from '../src/services/automower/accessTokenManager';
+import { AccessTokenManager } from '../src/services/husqvarna/accessTokenManager';
 import { DiscoveryService } from '../src/services/automower/discoveryService';
 import { EventStreamService } from '../src/services/automower/eventStreamService';
