+++ test/services/automower/eventStreamServiceImpl.spec.ts
@@ -5,7 +5,7 @@ import { BadCredentialsError } from '../../../src/errors/badCredentialsError';
 import { AutomowerEventTypes, PositionsEvent, SettingsEvent, StatusEvent } from '../../../src/events';
 import { AccessToken, Activity, HeadlightMode, Mode, OverrideAction, RestrictedReason, State } from '../../../src/model';
 import { Timer } from '../../../src/primitives/timer';
-import { AccessTokenManager } from '../../../src/services/automower/accessTokenManager';
+import { AccessTokenManager } from '../../../src/services/husqvarna/accessTokenManager';
 import { AutomowerEventStreamClientStub } from '../../clients/automowerEventStreamClientStub';
 import { EventStreamServiceImplSpy } from './eventStreamServiceImplSpy';
