+++ test/services/automower/getMowersServiceImpl.spec.ts
@@ -3,7 +3,7 @@ import { Mock, Times } from 'moq.ts';
 import { AutomowerClient } from '../../../src/clients/automowerClient';
 import { NotAuthorizedError } from '../../../src/errors/notAuthorizedError';
 import { AccessToken, Activity, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../src/model';
-import { AccessTokenManager } from '../../../src/services/automower/accessTokenManager';
+import { AccessTokenManager } from '../../../src/services/husqvarna/accessTokenManager';
 import { GetMowersServiceImpl } from '../../../src/services/automower/getMowersService';
 
 describe('GetMowersServiceImpl', () => {
